+++ duckscript/src/expansion_test.rs
@@ -39,6 +39,18 @@ fn expand_by_wrapper_only_control_chars() {
 }
 
 #[test]
+fn expand_by_wrapper_only_control_chars2() {
+    let mut variables = HashMap::new();
+    variables.insert("FOUND".to_string(), r#"\\\\"#.to_string());
+
+    let output = expand_by_wrapper("${FOUND}", &InstructionMetaInfo::new(), &mut variables);
+
+    let value = get_single_value(output);
+
+    assert_eq!(r#"\\\\"#, value);
+}
+
+#[test]
 fn expand_by_wrapper_found_fully() {
     let mut variables = HashMap::new();
     variables.insert("FOUND1".to_string(), "test1".to_string());